Return Ticker to libraries only for modularity.
